main_service/requirements.txt
fastapi[standard]==0.115.4
fast-depends==2.4.12
typing==3.7.4.3
pydantic==2.9.2
SQLAlchemy==2.0.36
uvicorn==0.32.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
httpx==0.27.2
python-logstash==0.4.8